Huge amounts of code fixing to fix problems with split in MTree.

+++ src/test/java/uk/ac/standrews/cs/utilities/m_tree/MTreeEuclidean2DTest.java
@@ -256,8 +256,8 @@ public class MTreeEuclidean2DTest {
         Point p = new Point(0.0F, 0.0F);
 
         // test search in ever increasing circles.
-        for (double i = 1.0F; i < 50.0F; i++) {
-            double search_circle = (double) Math.sqrt(i * i); // requested_result_set_size of square plus a little to avoid double errors
+        for (float i = 1.0F; i < 50.0F; i++) {
+            float search_circle = (float) Math.sqrt(i * i); // requested_result_set_size of square plus a little to avoid double errors
             List<DataDistance<Point>> result = tree.rangeSearch(p, search_circle);
             List<Point> values = tree.mapValues(result);
